+++ frontend/riskofbias/constants.js
@@ -21,6 +21,9 @@ const NA_KEYS = [10, 20],
         35: "-",
         36: "+",
         37: "++",
+
+        40: "Y",
+        41: "N",
     },
     SCORE_SHADES = {
         10: "#E8E8E8",
@@ -41,6 +44,9 @@ const NA_KEYS = [10, 20],
         35: "#FFCC00",
         36: "#6FFF00",
         37: "#00CC00",
+
+        40: "#00CC00",
+        41: "#CC3333",
     },
     SCORE_BAR_WIDTH_PERCENTAGE = {
         10: 50,
@@ -61,6 +67,9 @@ const NA_KEYS = [10, 20],
         35: 50,
         36: 75,
         37: 100,
+
+        40: 100,
+        41: 25,
     },
     SCORE_TEXT_DESCRIPTION = {
         10: "Not applicable",
@@ -81,6 +90,9 @@ const NA_KEYS = [10, 20],
         35: "Low confidence",
         36: "Medium confidence",
         37: "High confidence",
+
+        40: "Yes",
+        41: "No",
     },
     SCORE_TEXT_DESCRIPTION_LEGEND = {
         10: "Not applicable",
@@ -96,6 +108,9 @@ const NA_KEYS = [10, 20],
         25: "Deficient (metric) or Low confidence (overall)",
         26: "Adequate (metric) or Medium confidence (overall)",
         27: "Good (metric) or High confidence (overall)",
+
+        40: "Yes",
+        41: "No",
     },
     FOOTNOTES = {
         MULTIPLE_SCORES: ["✱", "Multiple judgments exist"],
