+++ hawc/apps/riskofbias/fixtures/ohat_study_quality_defaults.json
@@ -9,19 +9,22 @@
           "name": "Was administered dose or exposure level adequately randomized?",
           "description": "Randomization requires that each human subject or animal had an equal chance of being assigned to any study group including controls (e.g., use of random number table or computer generated randomization).",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         },
         {
           "name": "Was allocation to study groups adequately concealed?",
           "description": "Allocation concealment requires that research personnel do not know which administered dose or exposure level is assigned at the start of a study. Human studies also require that allocation be concealed from human subjects prior to entering the study. <br><br>Note: <br><ol><li>a question under performance bias addresses blinding of personnel and human subjects to treatment during the study;</li><li>a question under detection bias addresses blinding of outcome assessors.</li></ol>",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         },
         {
           "name": "Did selection of study participants result in appropriate comparison groups?",
           "description": "Comparison group appropriateness refers to having similar baseline characteristics between the groups aside from the exposures and outcomes under study.",
           "required_animal": false,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -33,7 +36,8 @@
           "name": "Did the study design or analysis account for important confounding and modifying variables?",
           "description": "",
           "required_animal": false,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -45,13 +49,15 @@
           "name": "Were experimental conditions identical across study groups?",
           "description": "",
           "required_animal": true,
-          "required_epi": false
+          "required_epi": false,
+          "responses": 1
         },
         {
           "name": "Were the research personnel and human subjects blinded to the study group during the study?",
           "description": "Blinding requires that study scientists do not know which administered dose or exposure level the human subject or animal is being given (i.e., study group). Human studies also require blinding of the human subjects when possible.",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -63,7 +69,8 @@
           "name": "Were outcome data incomplete due to attrition or exclusion from analysis?",
           "description": "Attrition rates are required to be similar and uniformly low across groups with respect to withdrawal or exclusion from analysis",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -75,13 +82,15 @@
           "name": "Can we be confident in the exposure characterization?",
           "description": "Confidence requires valid, reliable, and sensitive methods to measure exposure applied consistently across groups.",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         },
         {
           "name": "Can we be confident in the outcome assessment?",
           "description": "Confidence requires valid, reliable, and sensitive methods to assess the outcome and the methods should be applied consistently across groups.",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -93,7 +102,8 @@
           "name": "Were all measured outcomes reported?",
           "description": "",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     },
@@ -105,7 +115,8 @@
           "name": "Were there any other potential threats to internal validity?",
           "description": "<p>On a project specific basis, additional questions for other potential threats to internal validity can be added and applied to study designs as appropriate.</p><br><p>Examples may include appropriateness of statistical methods, adherence to the study-protocol, etc.</p>",
           "required_animal": true,
-          "required_epi": true
+          "required_epi": true,
+          "responses": 1
         }
       ]
     }
