+++ hawc/apps/riskofbias/models.py
@@ -17,7 +17,7 @@ from ..assessment.models import Assessment
 from ..common.helper import HAWCDjangoJSONEncoder, SerializerHelper, cleanHTML
 from ..myuser.models import HAWCUser
 from ..study.models import Study
-from . import managers
+from . import constants, managers
 
 logger = logging.getLogger(__name__)
 
@@ -70,31 +70,6 @@ class RiskOfBiasDomain(models.Model):
             child.copy_across_assessments(cw)
 
 
-class RiskOfBiasResponses(models.IntegerChoices):
-    NONE = 0, "None"
-    HIGH_LOW_BIAS = 1, "High/Low risk of bias"
-    GOOD_DEFICIENT = 2, "Good/deficient"
-    HIGH_LOW_CONFIDENCE = 3, "High/low confidence"
-    YES_NO = 4, "Yes/No"
-
-
-RESPONSES_VALUES = {
-    RiskOfBiasResponses.NONE: [0],
-    RiskOfBiasResponses.HIGH_LOW_BIAS: [17, 16, 15, 12, 14, 10],
-    RiskOfBiasResponses.GOOD_DEFICIENT: [27, 26, 25, 24, 22, 20],
-    RiskOfBiasResponses.HIGH_LOW_CONFIDENCE: [37, 36, 35, 34, 22, 20],
-    RiskOfBiasResponses.YES_NO: [40, 41, 22, 20],
-}
-
-RESPONSES_VALUES_DEFAULT = {
-    RiskOfBiasResponses.NONE: 0,
-    RiskOfBiasResponses.HIGH_LOW_BIAS: 12,
-    RiskOfBiasResponses.GOOD_DEFICIENT: 22,
-    RiskOfBiasResponses.HIGH_LOW_CONFIDENCE: 22,
-    RiskOfBiasResponses.YES_NO: 22,
-}
-
-
 class RiskOfBiasMetric(models.Model):
     objects = managers.RiskOfBiasMetricManager()
 
@@ -129,7 +104,7 @@ class RiskOfBiasMetric(models.Model):
         verbose_name="Use the short name?",
         help_text="Use the short name in visualizations?",
     )
-    responses = models.PositiveSmallIntegerField(choices=RiskOfBiasResponses.choices)
+    responses = models.PositiveSmallIntegerField(choices=constants.RiskOfBiasResponses.choices)
     sort_order = models.PositiveSmallIntegerField(validators=[MinValueValidator(1)])
     created = models.DateTimeField(auto_now_add=True)
     last_updated = models.DateTimeField(auto_now=True)
@@ -154,7 +129,7 @@ class RiskOfBiasMetric(models.Model):
         return reverse("riskofbias:arob_detail", args=(self.domain.assessment_id,))
 
     def get_response_values(self):
-        return RESPONSES_VALUES[self.responses]
+        return constants.RESPONSES_VALUES[self.responses]
 
     def copy_across_assessments(self, cw):
         old_id = self.id
@@ -257,7 +232,7 @@ class RiskOfBias(models.Model):
             [
                 len(strip_tags(score.notes)) > 0
                 for score in self.scores.all()
-                if score.score not in RiskOfBiasScore.NA_SCORES
+                if score.score not in constants.NA_SCORES
             ]
         )
 
@@ -363,85 +338,6 @@ class RiskOfBias(models.Model):
 class RiskOfBiasScore(models.Model):
     objects = managers.RiskOfBiasScoreManager()
 
-    RISK_OF_BIAS_SCORE_CHOICES = (
-        (0, "None"),
-        (10, "Not applicable"),
-        (12, "Not reported"),
-        (14, "Definitely high risk of bias"),
-        (15, "Probably high risk of bias"),
-        (16, "Probably low risk of bias"),
-        (17, "Definitely low risk of bias"),
-        (20, "Not applicable"),
-        (22, "Not reported"),
-        (24, "Critically deficient"),
-        (25, "Deficient"),
-        (26, "Adequate"),
-        (27, "Good"),
-        (34, "Uninformative"),
-        (35, "Low confidence"),
-        (36, "Medium confidence"),
-        (37, "High confidence"),
-        (40, "Yes"),
-        (41, "No"),
-    )
-
-    RISK_OF_BIAS_SCORE_CHOICES_MAP = {k: v for k, v in RISK_OF_BIAS_SCORE_CHOICES}
-
-    NA_SCORES = (10, 20)
-
-    SCORE_SYMBOLS = {
-        0: "N/A",
-        10: "N/A",
-        12: "NR",
-        14: "--",
-        15: "-",
-        16: "+",
-        17: "++",
-        20: "N/A",
-        22: "NR",
-        24: "--",
-        25: "-",
-        26: "+",
-        27: "++",
-        34: "--",
-        35: "-",
-        36: "+",
-        37: "++",
-        40: "Y",
-        41: "N",
-    }
-
-    SCORE_SHADES = {
-        0: "#E8E8E8",
-        10: "#E8E8E8",
-        12: "#FFCC00",
-        14: "#CC3333",
-        15: "#FFCC00",
-        16: "#6FFF00",
-        17: "#00CC00",
-        20: "#E8E8E8",
-        22: "#FFCC00",
-        24: "#CC3333",
-        25: "#FFCC00",
-        26: "#6FFF00",
-        27: "#00CC00",
-        34: "#CC3333",
-        35: "#FFCC00",
-        36: "#6FFF00",
-        37: "#00CC00",
-        40: "#00CC00",
-        41: "#CC3333",
-    }
-
-    BIAS_DIRECTION_UNKNOWN = 0
-    BIAS_DIRECTION_UP = 1
-    BIAS_DIRECTION_DOWN = 2
-    BIAS_DIRECTION_CHOICES = (
-        (BIAS_DIRECTION_UNKNOWN, "not entered/unknown"),
-        (BIAS_DIRECTION_UP, "⬆ (away from null)"),
-        (BIAS_DIRECTION_DOWN, "⬇ (towards null)"),
-    )
-
     TEXT_CLEANUP_FIELDS = (
         "score",
         "notes",
@@ -451,10 +347,10 @@ class RiskOfBiasScore(models.Model):
     metric = models.ForeignKey(RiskOfBiasMetric, on_delete=models.CASCADE, related_name="scores")
     is_default = models.BooleanField(default=True)
     label = models.CharField(max_length=128, blank=True)
-    score = models.PositiveSmallIntegerField(choices=RISK_OF_BIAS_SCORE_CHOICES, default=0)
+    score = models.PositiveSmallIntegerField(choices=constants.SCORE_CHOICES, default=0)
     bias_direction = models.PositiveSmallIntegerField(
-        choices=BIAS_DIRECTION_CHOICES,
-        default=BIAS_DIRECTION_UNKNOWN,
+        choices=constants.BiasDirections.choices,
+        default=constants.BiasDirections.BIAS_DIRECTION_UNKNOWN,
         help_text="Judgment of direction of bias (⬆ = away from null, ⬇ = towards null); only add entry if important to show in visuals",
     )
     notes = models.TextField(blank=True)
@@ -518,11 +414,11 @@ class RiskOfBiasScore(models.Model):
 
     @property
     def score_symbol(self):
-        return self.SCORE_SYMBOLS[self.score]
+        return constants.SCORE_SYMBOLS[self.score]
 
     @property
     def score_shade(self):
-        return self.SCORE_SHADES[self.score]
+        return constants.SCORE_SHADES[self.score]
 
     @classmethod
     def delete_caches(cls, ids):
